feat: :sparkles: adds VM ID variable
This adds the VM ID variable to the debian-13-vm module, so that the VM ID can be specified directly. It was just quite annoying that the VM ID of the created VMs did not match the order of the VMs specified so that sldocker01h had ID 102 while sldocker02h had ID 101.