feat: :technologist: improves proxmox provider authentication
This improves authentication handling with the proxmox provider, allowing to specify either password or api token for authentication. Also this allows to specify the SSH private key and the SSH port for the connection.

Note that the pve_username variable must be specified for password and API authentication as it will also be used for the SSH connection.